+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## [2.1.2]
 
 - Ride action not working correctly on some furniture mod
+- Entity custom name visible if not character in playback